Birthdate Handling Adjustment
Added in a patch for handling edge cases in the birthdate handler, invalidating all SINs priorly generated to conform with a different multiplier to push back the SIN equivalent of the Y2K bug to February 4, 3886.  By then either the world of Shadowrun will be post-SIN (as SINs replaced SSNs) or we will be in a new age and only the immortal elves will give a rat's ass about it.

Altered the testing suite code (commented out) to reflect these changes.